生成ルール直し デモ用の修正（textfieldがつぶれる現象解決）
git-svn-id: file:///Users/kunishi/tmp/svnroot/mk_exam/trunk@23 6b580622-3a88-449c-acd5-1afada95e41a

+++ other/C source/sample_h.c
@@ -17,7 +17,7 @@ int main (void){
 		c = c+d;
 	}
 
-	e=c;
+	e=(int)c;
 
 	f=e+c;
